requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.13
h11==0.16.0
idna==3.10
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
passlib==1.7.4
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-mongodb>=0.7.0
langchain-google-genai>=2.0.0